requirements.txt
Flask==3.1.3
NLTK==3.8.1
numpy==1.24.3
scikit-learn==1.3.0
snowflake-connector-python==3.0.4
gunicorn==25.1.0
